+++ CHANGELOG.md
@@ -3,9 +3,10 @@
 
 1. [](#new)
     * Added `FormTrait::getAllFlashes()` method to get all the available form flash objects for the form
+    * Added timestamp to `FormFlash` objects
 1. [](#bugfix)
     * Fixed error in `ImageMedium::url()` if the image cache folder does not exist
-    * Fixed empty form flash name after update
+    * Fixed empty form flash name after file upload or form state update
     * Fixed a bug in `Route::withParam()` method
 
 # v1.6.10